linux5.15: add patch to fix bluetooth dongle
The patch is taken verbatim from this particular comment:

    https://gist.github.com/nevack/6b36b82d715dc025163d9e9124840a07#gistcomment-3818325

Note that there are several fake variants of this CSR dongle
(usb id 0a12:0001) that need different workarounds. The gist itself
mentions a workaround (force-suspend) that is already included in 5.15
and fixes some dongles but not mine.

The issue is also mentioned in kernel bugzilla at

    https://bugzilla.kernel.org/show_bug.cgi?id=60824#c187

Starting at comment 187 there are several reports of dongles fixed by
that patch (or a variation / rediscovery of the workaround).

https://github.com/void-linux/void-packages/commit/625fd0cf938aaed065d8086b28c7e7bb80aba1b0