void-artwork: update to 20220303.
Adds an svg version of the logo, a script to generate icons from it,
and generated icon files for the logo (with and without the VOID text)
at various sizes, which are installed to the hicolor (fallback) icon
theme.

The logo files are pre-generated to avoid makedep on librsvg-utils,
and they are small enough and change so rarely that it is not a big
deal to include them in git.

The source svg is slightly different than the one on alpha.de; it has
a small white border added to help with legibility on dark backgrounds.

This is a follow-up to https://github.com/void-linux/void-packages/pull/35604

https://github.com/void-linux/void-packages/commit/f578bc2984db3c9afddc992100388e4e6f9840d0