base-files: add plugdev group
Plugdev is used by many packages' udev rules for pluggable devices.
Currently, some of those packages use the system_groups trigger to add it,
some forget to add it, and others patch it to "users" or "uaccess"
(which breaks the packages on systems that don't use logind).

By adding plugdev to the default groups, we get rid of those problems.

https://github.com/void-linux/void-packages/commit/bcd498943006dd780aa127b91e1b79c446a69d5a