common: unset CMAKE_GENERATOR before sourcing templates
When building dependencies of packages this has to be unset to avoid
using e.g. CMAKE_GENERATOR="Unix Makefiles" for a package which expected
to build with the default of make_cmd=ninja, and would subsequently fail
with "ninja: error: loading 'build.ninja': No such file or directory"
during do_build() if do_configure() generated a Makefile instead of a
build.ninja file.

https://github.com/void-linux/void-packages/commit/9f7f5629c61ce55eb77495775abf8c77d056548d