boost: fix aligned_alloc brokenness for real
The previous patch does not work; I could reproduce that
libtorrent was still using the wrong stuff on my system and
segfaulting. Even if it did work, it would be fragile, since
there is no guarantee features.h will define _XOPEN_SOURCE (e.g.
defining _BSD_SOURCE or _DEFAULT_SOURCE will disable the lines
that define it).

Therefore, stop trying to be clever and simply make POSIX the
fallback behavior. It would be okay to drop all the system
specific logic too, since our boost build is for a single
target, but those are harmless and would just make for an
unnecessarily bigger patch.

https://github.com/void-linux/void-packages/commit/5b1be772c792a63090a07e4ad84e428cd87247d5