xbps: revbump for glibc 2.36
This is a workaround to make sure glibc gets updated before other
packages.

dash was updated recently and it is earlier in the alphabetical order compared
to glibc, so it got updated sooner than diffutils which has an INSTALL shell
script which fails because the shell interpreter - dash expects the
newer glibc version than what's available at that time.

https://github.com/void-linux/void-packages/commit/576382c2bbe3e72467842ee36b3e7264d49c2e8c