common/xbps-src/shutils/chroot.sh: support importing other keys
When packages in `hostdir/binpkgs` are signed, it causes `xbps-src` to
annoyingly request to import the associated key (every time, in the case
of automation tools like `xxbuild` or `ymir`). This allows the user to
add their personal public key `.plist` to `etc/repo-keys/` so this
prompting does not happen.

https://github.com/void-linux/void-packages/commit/9f79d71beeb592d1c9896a06d83d631c76b0a0ac