mdocml: fix section name bugs in man-cgi
because of the previous patch that allows man-cgi to use
`.gz`-compressed manpages, there were some issues in various places.

also adds patch that allows for injecting HTML into `<head>`, so we can
drop putting custom stuff in mandoc.css and just add a second stylesheet
when building the containers in void-infrastructure.

https://github.com/void-linux/void-packages/commit/45756e8cdef1556e125f0a1120decc0e54690738