common/hooks/pre-pkg/04-generate-runtime-deps: partially revert sorting
only sort shlib-deps, as those are automatically found.
regular dependencies should not be sorted as that can mess with the
order for things like virtual packages or alternatives groups, notably
for gawk (needs to be before anything that depends on the awk virtual
package) and dash (needs to be before bash to select the correct sh
alternatives group).

partial revert of 99c13999573f48339d8d17e29665ab30355516fe

https://github.com/void-linux/void-packages/commit/5e9fb8708cb3b7bd70ab069534fba16521bc4ded